Refactoring tests

+++ sync/sync.go
@@ -1,13 +1,13 @@
 package sync
 
 type Counter struct {
-	value int64
+	value int
 }
 
 func (c *Counter) Inc() {
 	c.value++
 }
 
-func (c *Counter) Value() int64 {
+func (c *Counter) Value() int {
 	return c.value
 }
